+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 3.5.3
 
-- New typed exceptions: `DuplicateKeyException`, `ForeignKeyViolationException`. [#416](https://github.com/isoos/postgresql-dart/pull/416) by [hurrba](https://github.com/hurrba)
+- New typed exceptions: `UniqueViolationException`, `ForeignKeyViolationException`. [#416](https://github.com/isoos/postgresql-dart/pull/416) by [hurrba](https://github.com/hurrba)
 
 ## 3.5.2